[chore](build) Split the compliation units to build them in parallel
