+++ be/src/vec/functions/simple_function_factory.h
@@ -61,6 +61,7 @@ void register_function_in(SimpleFunctionFactory& factory);
 void register_function_if(SimpleFunctionFactory& factory);
 void register_function_nullif(SimpleFunctionFactory& factory);
 void register_function_date_time_computation(SimpleFunctionFactory& factory);
+void register_function_date_time_computation_v2(SimpleFunctionFactory& factory);
 void register_function_timestamp(SimpleFunctionFactory& factory);
 void register_function_utility(SimpleFunctionFactory& factory);
 void register_function_json(SimpleFunctionFactory& factory);
@@ -199,6 +200,7 @@ public:
             register_function_if(instance);
             register_function_nullif(instance);
             register_function_date_time_computation(instance);
+            register_function_date_time_computation_v2(instance);
             register_function_timestamp(instance);
             register_function_utility(instance);
             register_function_date_time_to_string(instance);
